requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pyyaml
jinja2
aiofiles
opencv-python-headless
numpy
tqdm